requirements.txt
Flask
firebase-admin
gunicorn
google-generativeai
python-dotenv